image working , modifying video to float16 to fit weights into 1/7gb instead of 3.5 gb so that model infeerence gets enuf space on my 5.5gb GPURAM , and CPU ooffloading to load frames in parts to the model and not all at once to ensure GPU does not overflow in memory , saving masks to  ouputs folder and docs havee explanation and how to revert to original infeerence method using flags

+++ sam3/model/model_misc.py
@@ -273,7 +273,7 @@ def gen_sineembed_for_position(pos_tensor, num_feats=256):
         pos = torch.cat((pos_y, pos_x, pos_w, pos_h), dim=2)
     else:
         raise ValueError("Unknown pos_tensor shape(-1):{}".format(pos_tensor.size(-1)))
-    return pos
+    return pos.to(pos_tensor.dtype)
 
 
 class SAM3Output(list):
